requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.10
django-decouple==2.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
pillow==10.2.0
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
stripe==8.4.0
typing-extensions==4.10.0
urllib3==2.2.1
